runtime: use i64 for watchdog timeout, not i32.

+++ artiq/test/libartiq_support/artiq_time.c
@@ -3,7 +3,7 @@
 
 int64_t now = 0;
 
-int watchdog_set(int ms)
+int watchdog_set(long long ms)
 {
   printf("watchdog_set %d\n", ms);
   return ms;
